Fix warnings

+++ LibraryDX12/Resource/VertexOrIndexBuffer.h
@@ -12,7 +12,7 @@ public:
     VertexOrIndexBuffer(ID3D12DevicePtr device, CommandList &commandList, const void *data, UINT size)
         : Resource(device, D3D12_HEAP_TYPE_DEFAULT, D3D12_HEAP_FLAG_NONE, size, D3D12_RESOURCE_STATE_COPY_DEST, nullptr) {
         const auto desc = this->resource->GetDesc();
-        recordGpuUploadCommands(device, commandList, data, desc.Width, desc.Height);
+        recordGpuUploadCommands(device, commandList, data, static_cast<UINT>(desc.Width), desc.Height);
     }
 };
